remove modules, move abilities and update it.
