+++ src/Parser/Druid/Balance/CombatLogParser.js
@@ -7,7 +7,7 @@ import GlobalCooldown from './Modules/Core/GlobalCooldown';
 import Checklist from './Modules/Features/Checklist';
 import AlwaysBeCasting from './Modules/Features/AlwaysBeCasting';
 import CancelledCasts from './Modules/Features/CancelledCasts';
-import Abilities from './Modules/Features/Abilities';
+import Abilities from './Modules/Abilities';
 import LunarEmpowerment from './Modules/Features/LunarEmpowerment';
 import SolarEmpowerment from './Modules/Features/SolarEmpowerment';
 import CooldownThroughputTracker from './Modules/Features/CooldownThroughputTracker';
@@ -15,16 +15,10 @@ import MoonfireUptime from './Modules/Features/MoonfireUptime';
 import SunfireUptime from './Modules/Features/SunfireUptime';
 import StellarFlareUptime from './Modules/Features/StellarFlareUptime';
 import StellarEmpowermentUptime from './Modules/Features/StellarEmpowermentUptime';
-import MoonSpells from './Modules/Features/MoonSpells';
 import UnempoweredLunarStrike from './Modules/Features/UnempoweredLunarStrike';
 import UnempoweredSolarWrath from './Modules/Features/UnempoweredSolarWrath';
 
 //Talents
-import L90Talents from './Modules/Talents/L90Talents';
-import StellarDriftMissing from './Modules/Talents/StellarDriftMissing';
-import StellarDrift from './Modules/Talents/StellarDrift';
-import StellarFlare from './Modules/Talents/StellarFlare';
-import Starlord from './Modules/Talents/Starlord';
 
 import EarlyDotRefreshes from './Modules/Features/EarlyDotRefreshes';
 import EarlyDotRefreshesInstants from './Modules/Features/EarlyDotRefreshesInstants';
@@ -64,16 +58,10 @@ class CombatLogParser extends MainCombatLogParser {
     sunfireUptime: SunfireUptime,
     stellarFlareUptime: StellarFlareUptime,
     stellarEmpowermentUptime: StellarEmpowermentUptime,
-    moonSpells: MoonSpells,
     unempoweredLunarStrike: UnempoweredLunarStrike,
     unempoweredSolarWrath: UnempoweredSolarWrath,
 
     //Talents
-    l90Talents: L90Talents,
-    stellarDriftMissing: StellarDriftMissing,
-    stellarDrift: StellarDrift,
-    stellarFlare: StellarFlare,
-    starlord: Starlord,
 
     earlyDotRefreshes: EarlyDotRefreshes,
     earlyDotRefreshesInstants: EarlyDotRefreshesInstants,
