--- src/Parser/Druid/Balance/Modules/Core/isMoonMoon.js
@@ -1,5 +1,0 @@
-import SPELLS from 'common/SPELLS';
-
-export default function isMoonMoon(event) {
-  return [SPELLS.FULL_MOON.id, SPELLS.HALF_MOON.id, SPELLS.NEW_MOON.id].includes(event.ability.guid);
-}
